fixed memory leak

+++ Assets/Scripts/Interfaces/IWorldHelper.cs
@@ -14,14 +14,14 @@ public interface IWorldHelper
     
     public Vector3Int Vec3ToBlockPos(Vector3 pos);
 
-     public void SetBlockByHand(Vector3 pos,int blockID);
-         public void SetBlock(Vector3 pos,int blockID);
-         public void SetBlockWithoutUpdate(Vector3 pos,int blockID);
+     public void SetBlockByHand(Vector3 pos,short blockID);
+         public void SetBlock(Vector3 pos,short blockID);
+         public void SetBlockWithoutUpdate(Vector3 pos,short blockID);
     public int GetChunkLandingPoint(float x, float z);
   
-    public int GetBlock(Vector3 pos);
-    public int GetBlock(Vector3 pos,Chunk chunkNeededUpdate);
+    public short GetBlock(Vector3 pos);
+    public short GetBlock(Vector3 pos,Chunk chunkNeededUpdate);
 
-    public int GetBlockInSingleChunk(Vector3 chunkSpacePos,Chunk chunkNeededUpdate);
+    public short GetBlockInSingleChunk(Vector3 chunkSpacePos,Chunk chunkNeededUpdate);
 }
